+++ projects/ng-keeps-ui/src/lib/components/kp-filter/index.ts
@@ -1,0 +1,2 @@
+export * from './kp-filter.module';
+export * from './kp-filter.component';