+++ CHANGELOG.md
@@ -2,4 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.4](https://github.com/renato-rjps/keeps-angular-components/compare/v0.0.2...v0.0.4) (2020-04-17)
+
+
+### Features
+
+* **css:** add global css to lib kp-components ([1aaf49b](https://github.com/renato-rjps/keeps-angular-components/commit/1aaf49befe35767b7dbcd4bbefe9f3988f99ba60))
+* **ng-keeps-ui:** create lib ng-keeps-ui ([3eaaa69](https://github.com/renato-rjps/keeps-angular-components/commit/3eaaa69a239b07e2c3df4f44fa18cbdca20641de))
+
 ### [0.0.2](https://github.com/renato-rjps/keeps-angular-components/compare/v0.0.1...v0.0.2) (2020-04-16)